scripts/train.py
#!/usr/bin/env python3
"""
Loka Training Script

Train the Loka astrophysics navigation model.
"""

import argparse
import os
from pathlib import Path

import torch
import yaml
from accelerate import Accelerator
from datasets import load_dataset
from torch.utils.data import DataLoader
from transformers import (
    AutoModelForCausalLM,
    AutoTokenizer,
    TrainingArguments,
    Trainer,
    get_scheduler,
)

def parse_args():
    parser = argparse.ArgumentParser(description="Train Loka model")
    parser.add_argument(
        "--config",
        type=str,
        default="configs/train_config.yaml",
        help="Path to training config file",
    )
    parser.add_argument(
        "--output_dir",
        type=str,
        default=None,
        help="Output directory for model checkpoints",
    )
    parser.add_argument(
        "--resume_from",
        type=str,
        default=None,
        help="Resume training from checkpoint",
    )
    return parser.parse_args()


def load_config(config_path: str) -> dict:
    """Load training configuration from YAML file."""
    with open(config_path, "r") as f:
        config = yaml.safe_load(f)
    return config


def main():
    args = parse_args()
    config = load_config(args.config)
    
    # Initialize accelerator for distributed training
    accelerator = Accelerator(
        mixed_precision=config.get("mixed_precision", "fp16"),
        gradient_accumulation_steps=config.get("gradient_accumulation_steps", 4),
    )
    
    # Set output directory
    output_dir = args.output_dir or os.environ.get(
        "LOKA_MODEL_DIR", "./models"
    )
    output_dir = Path(output_dir)
    output_dir.mkdir(parents=True, exist_ok=True)
    
    accelerator.print(f"Training Loka model")
    accelerator.print(f"Output directory: {output_dir}")
    accelerator.print(f"Config: {config}")
    
    # Load tokenizer and model
    model_name = config.get("base_model", "mistralai/Mistral-7B-v0.1")
    accelerator.print(f"Loading base model: {model_name}")
    
    tokenizer = AutoTokenizer.from_pretrained(model_name)
    if tokenizer.pad_token is None:
        tokenizer.pad_token = tokenizer.eos_token
    
    model = AutoModelForCausalLM.from_pretrained(
        model_name,
        torch_dtype=torch.float16 if config.get("mixed_precision") == "fp16" else torch.float32,
        device_map="auto" if not accelerator.distributed_type else None,
    )
    
    # TODO: Load and preprocess astrophysics dataset
    # This would include:
    # - Celestial body positions and ephemeris data
    # - Trajectory planning examples
    # - Mission planning dialogues
    # - Orbital mechanics calculations
    
    accelerator.print("Loading training dataset...")
    # Placeholder: Replace with actual dataset loading
    # dataset = load_dataset("path/to/loka_dataset")
    
    # Training arguments
    training_args = TrainingArguments(
        output_dir=str(output_dir),
        num_train_epochs=config.get("num_epochs", 3),
        per_device_train_batch_size=config.get("batch_size", 4),
        gradient_accumulation_steps=config.get("gradient_accumulation_steps", 4),
        learning_rate=float(config.get("learning_rate", 2e-5)),
        warmup_ratio=config.get("warmup_ratio", 0.1),
        weight_decay=config.get("weight_decay", 0.01),
        logging_steps=config.get("logging_steps", 10),
        save_steps=config.get("save_steps", 500),
        save_total_limit=config.get("save_total_limit", 3),
        fp16=config.get("mixed_precision") == "fp16",
        bf16=config.get("mixed_precision") == "bf16",
        dataloader_num_workers=config.get("num_workers", 4),
        report_to=config.get("report_to", ["wandb"]),
        run_name=config.get("run_name", "loka-training"),
    )
    
    accelerator.print("Training arguments configured")
    accelerator.print(f"Training will use {accelerator.num_processes} processes")
    
    # TODO: Initialize trainer with actual dataset
    # trainer = Trainer(
    #     model=model,
    #     args=training_args,
    #     train_dataset=dataset["train"],
    #     eval_dataset=dataset.get("validation"),
    #     tokenizer=tokenizer,
    # )
    
    # trainer.train(resume_from_checkpoint=args.resume_from)
    
    accelerator.print("Training setup complete. Dataset loading not yet implemented.")
    accelerator.print("See src/loka/data/ for dataset preparation utilities.")


if __name__ == "__main__":
    main()
